moved extract function
